requirements.txt
psycopg2
qdrant-client
openai==0.27.8
openai[embeddings]
tiktoken
django==4.1.5
django.apps
djangorestframework==3.14.0
dj_database_url
gunicorn
whitenoise
pinecone-client
requests
jsonschema